Add unit test for #311.
Problem: We have an issue (#311) that causes a crash but we don't have a
unit test for it.

Solution: This adds a unit test that also crashes for what is believed
to be the same cause.  This commit should actually be squashed with the
bugfix (once we have it) so that we have a bugfix and a test.